+++ src/metrics.cpp
@@ -1,0 +1,10 @@
+/**
+    metrics.cpp
+    Purpose: Mesure Behaviour of Turing Machines.
+
+    @author Jorge Miguel Ferreira da Silva
+    @version 0.1
+*/
+
+#include "metrics.h"
+
